tests: Move SewMessageAndReactionTest to test_message_dict.py.
